Split cgroup.py and rename classes
