+++ azurelinuxagent/common/exception.py
@@ -20,6 +20,13 @@
 Defines all exceptions
 """
 
+class CGroupsException(Exception):
+
+    def __init__(self, msg):
+        self.msg = msg
+
+    def __str__(self):
+        return repr(self.msg)
 
 class AgentError(Exception):
     """
